moved server to typescript and common to server
