+++ client/tsconfig.json
@@ -28,8 +28,5 @@
     "strictInjectionParameters": true,
     "strictInputAccessModifiers": true,
     "strictTemplates": true
-  },
-  "paths": {
-    "common": ["../common"]
   }
 }
