--- common/package.json
@@ -1,6 +1,0 @@
-{
-  "name": "common",
-  "version": "0.0.0",
-  "description": "common interfaces between the super-bowl-quiz client and server",
-  "main": "interfaces.ts"
-}
