+++ server/interfaces.ts
