+++ server/package.json
@@ -1,10 +1,11 @@
 {
-  "name": "superbowlquiz",
-  "version": "1.0.0",
+  "name": "super-bowl-quiz-server",
+  "version": "0.0.0",
   "description": "Super Bowl Quiz for the 2020 Super Bowl",
   "main": "index.ts",
   "scripts": {
     "start": "ts-node index.ts",
+    "dev": "nodemon",
     "admin-console": "ts-node admin-console.ts"
   },
   "keywords": [
@@ -15,14 +16,15 @@
   "author": "Joshua Fishman",
   "license": "MIT",
   "dependencies": {
+    "dotenv": "^16.0.3",
     "express": "^4.18.2",
-    "fs": "0.0.1-security",
-    "pg": "^8.5.1",
+    "pg": "^8.8.0",
     "readline": "^1.3.0"
   },
   "devDependencies": {
     "@types/express": "^4.17.15",
     "@types/node": "^18.11.17",
-    "@types/pg": "^8.6.6"
+    "@types/pg": "^8.6.6",
+    "nodemon": "^2.0.20"
   }
 }
